requirements.txt
torch>=2.2
transformers>=4.40
peft>=0.10
datasets>=2.19
accelerate>=0.27
bitsandbytes>=0.42
optimum-intel>=1.18
pydicom
pillow
gradio>=4.14
scikit-learn
rouge-score
evaluate>=0.4.0
